refactor: move Terminable
