+++ src/utils/disposable.ts
@@ -1,4 +1,20 @@
-import { Terminable } from '../core/encryption/processors/WorkerEncryptionProxyProcessor';
+export class Terminable {
+	private isFinished = false;
+
+	public terminate() {
+		this.isFinished = true;
+	}
+
+	public isTerminated() {
+		return this.isFinished;
+	}
+
+	public throwErrorIfTerminated(message?: string) {
+		const errorMessage =
+			message ?? "Object been terminated and can't be used anymore";
+		if (this.isFinished) throw new Error(errorMessage);
+	}
+}
 
 export class DisposableBox<T> {
 	private readonly data;
